Complete exercise #1 of Blocks and Margin CSS course
Body: More detailed explanatory text: wrap it to about 72 characters
or so (optional).Here you explain the problem that this commit is
solving. Focus on why you are making this change as opposed to how
(the code explains that). Are there side effects or other unintuitive
consequences of this change? Here's the place to explain them.

 - The hyphen is preceded by a single space, with blank lines in
 between, but conventions vary here.

At the bottom is the issue or bug tracket ticket tracking (optional).

Resolves: #XXX
See also: #XXX